Add additional detection for transitive deletes and fix those found
We were not testing for transitive deletes when performing *User*
deletion, which led us to have a few
